Requirement4

See merge request se-2022/group-55/EzWh!4

+++ code/server/classes/SKU/SKUAPI.js
@@ -27,7 +27,7 @@ module.exports = function(app){
     app.get('/api/skus/:id', async function (req, res) {
         try {
             //if(req.session.loggedin && req.session.user.type=="manager"){
-            if (req.params.id != undefined) {
+            if (req.params.id != undefined && !isNaN(parseInt(req.params.id))) {
                 const sku = await skudao.getSKUByID(req.params.id);
                 if (sku != null) {
                     sku.setTestDescriptorIDList(await skudao.getTestDescriptorsBySKUID(sku.id));
@@ -47,7 +47,14 @@ module.exports = function(app){
     app.post('/api/sku', async function (req, res) {
         try {
             //if(req.session.loggedin && req.session.user.type=="manager"){
-            if (req.body.description != undefined && req.body.weight != undefined && req.body.volume != undefined && req.body.notes != undefined && req.body.price != undefined && req.body.availableQuantity != undefined) {
+            if (req.body.id != undefined || req.body.description != undefined && req.body.weight != undefined && req.body.volume != undefined && req.body.notes != undefined && req.body.price != undefined && req.body.availableQuantity != undefined) {
+                if(!Number.isInteger(req.body.id) && req.body.description==="" || req.body.notes==="" || typeof req.body.weight!="number" || !Number.isInteger(req.body.weight) || typeof req.body.volume!="number" || !Number.isInteger(req.body.volume) || typeof req.body.price!="number" || typeof req.body.availableQuantity!="number" || !Number.isInteger(req.body.availableQuantity) || req.body.weight<0 || req.body.volume<0 || req.body.price<0 || req.body.availableQuantity<0){
+                    return res.status(422).end();
+                }
+                const sku = await skudao.getSKUByID(req.body.id);
+                if(sku!=null){
+                    return res.status(422).end();
+                }
                 await skudao.storeSKU(req.body);
                 return res.status(201).end();
             }
@@ -63,7 +70,7 @@ module.exports = function(app){
     app.put('/api/sku/:id', async function (req, res) {
         try {
             //if(req.session.loggedin && req.session.user.type=="manager"){
-            if (req.params.id != undefined && req.body.newDescription != undefined && req.body.newWeight != undefined && req.body.newVolume != undefined && req.body.newNotes != undefined && req.body.newPrice != undefined && req.body.newAvailableQuantity != undefined) {
+            if (req.params.id != undefined && !isNaN(parseInt(req.params.id)) && req.body.newDescription != undefined && req.body.newWeight != undefined && req.body.newVolume != undefined && req.body.newNotes != undefined && req.body.newPrice != undefined && req.body.newAvailableQuantity != undefined) {
                 const sku = await skudao.getSKUByID(req.params.id);
                 if (sku != null) {
                     if (sku.availableQuantity != req.body.newAvailableQuantity) {
@@ -96,7 +103,7 @@ module.exports = function(app){
     app.put('/api/sku/:id/position', async function (req, res) {
         try {
             //if(req.session.loggedin && req.session.user.type=="manager"){
-            if (req.params.id != undefined && req.body.position != undefined) {
+            if (req.params.id != undefined && !isNaN(parseInt(req.params.id)) && req.body.position != undefined) {
                 const sku = await skudao.getSKUByID(req.params.id)
                 if (sku != null) {
                     const weight = sku.availableQuantity * sku.weight;
@@ -135,15 +142,15 @@ module.exports = function(app){
     app.delete('/api/skus/:id', async function (req, res) {
         try {
             //if(req.session.loggedin && req.session.user.type=="manager"){
-            if (req.params.id != undefined) {
+            if (req.params.id != undefined && !isNaN(parseInt(req.params.id))) {
                 const sku = await skudao.getSKUByID(req.params.id);
                 if (sku != null) {
-                    if (await skudao.existingSKUItem(req.params.id)) {
+                   /* if (await skudao.existingSKUItem(req.params.id)) {
                         return res.status(422).end();
                     }
                     if (await skudao.existingTestDescriptor(req.params.id)) {
                         return res.status(422).end();
-                    }
+                    }*/
                     if (sku.position != null) {
                         await skudao.updatePositionWeightVolume(sku.position, 0, 0);
                         await skudao.deleteSKU(req.params.id);
@@ -153,7 +160,7 @@ module.exports = function(app){
                         return res.status(204).end();
                     }
                 }
-                return res.status(422).end();
+                return res.status(204).end();
             }
             return res.status(422).end();
             //}else{
